requirements.txt
ruff
python-dotenv
requests
gitpython
pandas
matplotlib
ujson